moving all the mixin classes to mixin package
