--- ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/deserialize/CompanyDetailsMixIn.java
@@ -1,6 +1,0 @@
-package uk.gov.hmcts.cmc.ccd.deserialize;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public abstract class CompanyDetailsMixIn extends CompanyMixIn {
-}
