--- ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/deserialize/IndividualDetailsMixIn.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.cmc.ccd.deserialize;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.time.LocalDate;
-
-public abstract class IndividualDetailsMixIn extends IndividualMixIn {
-}
