+++ ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/jackson/mixin/CompanyDetailsMixIn.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.cmc.ccd.jackson.mixin;
+
+public abstract class CompanyDetailsMixIn extends CompanyMixIn {
+}
