+++ ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/jackson/mixin/InterestBreakDownMixIn.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.ccd.deserialize;
+package uk.gov.hmcts.cmc.ccd.jackson.mixin;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
