+++ src/main/java/uk/gov/hmcts/cmc/claimstore/config/JacksonConfiguration.java
@@ -10,27 +10,26 @@ import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-import uk.gov.hmcts.cmc.ccd.deserialize.ClaimDataMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.CompanyDetailsMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.CompanyMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.EvidenceMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.IndividualDetailsMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.IndividualMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.InterestBreakDownMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.InterestDateMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.InterestMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.OrganisationDetailsMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.OrganisationMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.RepresentativeMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.SoleTraderDetailsMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.SoleTraderMixIn;
-import uk.gov.hmcts.cmc.ccd.deserialize.TimelineMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.ClaimDataMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.CompanyDetailsMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.CompanyMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.EvidenceMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.IndividualDetailsMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.IndividualMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.InterestBreakDownMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.InterestDateMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.InterestMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.OrganisationDetailsMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.OrganisationMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.RepresentativeMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.SoleTraderDetailsMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.SoleTraderMixIn;
+import uk.gov.hmcts.cmc.ccd.jackson.mixin.TimelineMixIn;
 import uk.gov.hmcts.cmc.domain.models.ClaimData;
 import uk.gov.hmcts.cmc.domain.models.Interest;
 import uk.gov.hmcts.cmc.domain.models.InterestBreakdown;
 import uk.gov.hmcts.cmc.domain.models.InterestDate;
 import uk.gov.hmcts.cmc.domain.models.Timeline;
-import uk.gov.hmcts.cmc.domain.models.amount.AmountBreakDown;
 import uk.gov.hmcts.cmc.domain.models.evidence.Evidence;
 import uk.gov.hmcts.cmc.domain.models.legalrep.Representative;
 import uk.gov.hmcts.cmc.domain.models.otherparty.CompanyDetails;
